updated Product Schema and product route to include sellerId

+++ routes/product.js
@@ -7,7 +7,10 @@ const router = Router();
 
 router.get("/products", async (req, res) => {
   try {
-    const products = await Product.find();
+    const products = await Product.find().populate(
+      "sellerId",
+      "-password -__v -createdAt -updatedAt"
+    );
     res.status(200).json(products);
   } catch (error) {
     res.status(500).json(`Something went wrong and an error occured: ${error}`);
@@ -16,7 +19,9 @@ router.get("/products", async (req, res) => {
 
 router.get("/products/:catogory", async (req, res) => {
   try {
-    const products = await Product.find({ category: req.params.catogory });
+    const products = await Product.find({
+      category: req.params.catogory,
+    }).populate("sellerId", "-password -__v -createdAt -updatedAt");
     res.status(200).json(products);
   } catch (error) {
     res.status(500).json(`Something went wrong and an error occured: ${error}`);
@@ -30,7 +35,7 @@ router.post("/products/create", verifyToken, async (req, res) => {
         .status(400)
         .json("You are not permitted to perform this operation");
 
-    const product = new Product(req.body);
+    const product = new Product({ ...req.body, sellerId: req.user._id });
     const createdProduct = await product.save();
     res.status(201).json(createdProduct);
   } catch (error) {
@@ -44,6 +49,10 @@ router.patch("/product/update/:id", verifyToken, async (req, res) => {
       return res
         .status(400)
         .json("You are not permitted to perform this operation");
+
+    const product = await Product.findOne({ _id: req.params.id });
+    if (product.sellerId.toString() !== req.user._id)
+      return res.status(400).json("Your are not the seller of this product!");
 
     const updatedProduct = await Product.findOneAndUpdate(
       { _id: req.params.id },
@@ -62,6 +71,10 @@ router.delete("/product/delete/:id", verifyToken, async (req, res) => {
       return res
         .status(400)
         .json("You are not permitted to perform this operation");
+
+    const product = await Product.findOne({ _id: req.params.id });
+    if (product.sellerId.toString() !== req.user._id)
+      return res.status(400).json("Your are not the seller of this product!");
 
     const deletedProduct = await Product.findOneAndDelete({
       _id: req.params.id,
